Begin fixing circular dependencies
