+++ src/commands/math.ts
@@ -5,7 +5,7 @@
 import { MQNode } from "../services/keystroke";
 import { CursorOptions, NodeRef, MathspeakOptions, CharCmdsAny, InnerFields, JoinMethod, LatexCmdsAny, LatexCmdsSingleCharBuilder } from "../shared_types";
 import { Anticursor, Cursor, MQSelection } from "../cursor"
-import { API, APIClasses, Options, RootBlockMixin } from "../publicapi"
+import { API, APIClasses, RootBlockMixin } from "../publicapi"
 import { Fragment, Ends, NodeBase, CharCmds, LatexCmds } from "../tree";
 import { Direction, pray, L, R, noop } from "../utils"
 import { Parser, } from "../services/parser.util";
@@ -15,6 +15,7 @@ import { getBoundingClientRect } from "../browser";
 import { Controller } from "../services/textarea";
 import { latexMathParser } from "src/services/latex";
 import { Digit, Letter } from "./math/basicSymbols";
+import { Options } from "src/options";
 
 /**
  * Math tree node base class.
