+++ src/services/textarea.ts
@@ -4,11 +4,11 @@
  ********************************************/
 
 import { h } from "../dom"
-import { Options } from "../publicapi"
 import { domFrag, jQToDOMFragment } from "../domFragment"
 import { Controller_scrollHoriz } from "./scrollHoriz"
 import { saneKeyboardEvents } from "./saneKeyboardEvents.util"
 import { MQNode } from "./keystroke"
+import { Options } from "src/options"
 
 Options.prototype.substituteTextarea = function () {
   return h('textarea', {
